requirements.txt
Django==5.1.1
pandas==2.2.3
openpyxl==3.1.5
qrcode[pil]==8.0
Pillow==10.4.0
python-decouple==3.8